+++ src/elements/Space.lua
@@ -6,11 +6,11 @@ local Element = {}
 function Element:New(Config)
     local MainSpace = New("Frame", {
         Parent = Config.Parent,
-        Size = UDim2.new(1,-7,0,7),
+        Size = UDim2.new(1,-7,0,7*(Config.Columns or 1)),
         BackgroundTransparency = 1,
     })
     
-    return "Space", { __type = "Space" }
+    return "Space", { __type = "Space", ElementFrame = MainSpace}
 end
 
 return Element