+++ src/utils/services/Luarmor.lua
@@ -10,7 +10,12 @@ local Luarmor = {}
 
 
 function Luarmor.New(scriptId, discord)
-    local API = loadstring(game:HttpGet("https://sdkapi-public.luarmor.net/library.lua"))()
+    local APIURL = "https://sdkapi-public.luarmor.net/library.lua"
+
+    local API = loadstring(
+        game.HttpGetAsync and game:HttpGetAsync(APIURL)
+        or HttpService:GetAsync(APIURL)
+    )()
     local fsetclipboard = setclipboard or toclipboard
 
     API.script_id = scriptId